#27 - better filenames, themed, removed collapse from system cards, ellipsis truncating, from review comments
